chore(common): rename kmx_process files
